services/temporal_service/requirements.txt
temporalio>=1.6.0
httpx>=0.25.0
pydantic>=2.0.0
PyYAML>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
